feat: use create password generation
